refactor(lib-sourcify): move getCompilerNameFromLanguage into PreRunCompilation
It was defined in AbstractCompilation but only PreRunCompilation reconstructs
a compiler name from a stored language; single-compiler compilations know
their compiler directly. Colocate it with its only consumer. Still exported
via the PreRunCompilation barrel, so the public API is unchanged.
