+++ packages/lib-sourcify/src/Compilation/AbstractCompilation.ts
@@ -26,20 +26,6 @@ import type {
 } from '@ethereum-sourcify/compilers-types';
 import { logDebug, logInfo, logSilly, logWarn } from '../logger';
 
-export function getCompilerNameFromLanguage(language: string): string {
-  switch (language.toLocaleLowerCase()) {
-    case 'yul':
-    case 'solidity':
-      return 'solc';
-    case 'vyper':
-      return 'vyper';
-    case 'fe':
-      return 'fe';
-    default:
-      throw new Error('Language not supported');
-  }
-}
-
 function cleanCompilerVersion(version: string): string {
   // Remove non-numerical characters from the beginning of the version string
   return version.replace(/^[^\d]*/, '');
@@ -82,9 +68,6 @@ export abstract class AbstractCompilation {
     forceEmscripten: boolean,
   ): Promise<SolidityOutput | VyperOutput | FeOutput>;
 
-  // The compiler name stored/exported for this compilation. Each compilation
-  // maps 1:1 to a compiler, so it is declared per class rather than inferred
-  // from the language (which is ambiguous, e.g. Solidity → solc or zksolc).
   public abstract get compilerName(): string;
 
   // The compiler version string stored/exported for this compilation. Defaults
